Integrando nuevas clases a las interfaces

+++ Vettrack/Mascotas/Ave.cs
@@ -6,6 +6,6 @@ namespace Vettrack.Mascotas
 {
     public class Ave
     {
-        int EnvergaduraAlas { get; set; }
+        double EnvergaduraAlas { get; set; }
     }
 }
